requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
